Added SignalR middleware to join groups
