+++ secrethitler.web/src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/lib/integration/react';
 import {configureStore, persistor} from './store/store';
-import { signalRRegistration } from './store/middleware/signalRmiddleware.js';
+import { signalRRegistration } from './store/middleware/signalRconnection.js';
 
 import './index.css';
 import App from './App';
