+++ secrethitler.web/src/store/store.js
@@ -9,6 +9,7 @@ import autoMergeLevel2 from 'redux-persist/lib/stateReconciler/autoMergeLevel2';
 
 import thunk from 'redux-thunk';
 import { apiMiddleware } from './middleware/apiMiddleware.js';
+import { signalRMiddleware } from './middleware/signalRMiddleware.js'
 import rootReducer from './reducers';
 
 const persistConfig = {
@@ -18,6 +19,6 @@ const persistConfig = {
 };
 
 const pReducer = persistReducer(persistConfig, rootReducer);
-
-export const configureStore = createStore(pReducer, applyMiddleware(thunk, apiMiddleware));
+//signalRmiddleware als derde middleware
+export const configureStore = createStore(pReducer, applyMiddleware(thunk, apiMiddleware, signalRMiddleware));
 export const persistor = persistStore(configureStore);
